modified the build procedure to honor the versioning of the boost-version it is built against

+++ contrib/boost.mpi-automake/1_35_0/bootstrap.sh
@@ -2,18 +2,20 @@
 
 BOOSTPATH=$1
 
-if test "x$BOOSTPATH" = "x" -o ! -f "$BOOSTPATH/LICENSE_1_0.txt"; then
+if test ! -f "LICENSE_1_0.txt" -a \( "x$BOOSTPATH" = "x" -o ! -f "$BOOSTPATH/LICENSE_1_0.txt" \) ; then
     echo "usage: ./boostrap.sh <path to boost source directory>"
     echo "this will extract the boost sources for Boost:MPI"
     exit -1
 fi
 
-# copy the files from boost
-mkdir -p ./boost
-for f in LICENSE_1_0.txt boost/mpi.hpp boost/mpi; do
-    cp -r $BOOSTPATH/$f ./$f;
-done
-cp -r $BOOSTPATH/libs/mpi/src .
+if test "x$BOOSTPATH" != "x"; then
+    # copy the files from boost
+    mkdir -p ./boost
+    for f in LICENSE_1_0.txt boost/mpi.hpp boost/mpi; do
+        cp -r $BOOSTPATH/$f ./$f;
+    done
+    cp -r $BOOSTPATH/libs/mpi/src .
+fi
 
 # and generate the autoconf/automake files
